Refactored vertical coordinate routines
