+++ mdtf_test_data/synthetic/synthetic_data.py
@@ -5,9 +5,6 @@ ___all__ = [
     "dataset_stats",
     "generate_synthetic_dataset",
     "generate_random_array",
-    "gfdl_vertical_coord",
-    "gfdl_plev19_vertical_coord",
-    "ncar_hybrid_coord",
     "write_to_netcdf",
 ]
 
@@ -20,6 +17,10 @@ import mdtf_test_data.generators as generators
 from mdtf_test_data.synthetic.time import generate_monthly_time_axis
 from mdtf_test_data.synthetic.time import generate_daily_time_axis
 from mdtf_test_data.synthetic.time import generate_hourly_time_axis
+
+from mdtf_test_data.synthetic.vertical import gfdl_plev19_vertical_coord
+from mdtf_test_data.synthetic.vertical import gfdl_vertical_coord
+from mdtf_test_data.synthetic.vertical import ncar_hybrid_coord
 
 
 def dataset_stats(filename, var=None, limit=None):
@@ -215,399 +216,6 @@ def generate_synthetic_dataset(
         dset.attrs = {**dset.attrs, **cmip_global_atts}
 
     return dset
-
-
-def gfdl_vertical_coord():
-    """Generates GFDL AM4 pressure coordinate
-
-    Returns
-    -------
-    xarray.DataArray
-        GFDL AM4 pressure levels and half levels
-    """
-
-    pfull = np.array(
-        [
-            2.164043,
-            5.845308,
-            10.74508,
-            17.106537,
-            25.113805,
-            35.221197,
-            48.137904,
-            64.560184,
-            85.114482,
-            110.419627,
-            141.09261,
-            177.729388,
-            220.892397,
-            271.066624,
-            328.516337,
-            392.785273,
-            461.947262,
-            532.465907,
-            600.430867,
-            663.107383,
-            719.307118,
-            768.814284,
-            811.846869,
-            848.836021,
-            880.346139,
-            906.995722,
-            929.394583,
-            948.128523,
-            963.73257,
-            976.687397,
-            987.392458,
-            996.109949,
-        ]
-    )
-
-    phalf = np.array(
-        [
-            1.0,
-            4.0,
-            8.186021,
-            13.788865,
-            20.917952,
-            29.836408,
-            41.217896,
-            55.792215,
-            74.201906,
-            97.047864,
-            124.966648,
-            158.549553,
-            198.396959,
-            245.027221,
-            298.888576,
-            360.040179,
-            427.458025,
-            498.243573,
-            568.220535,
-            633.836047,
-            693.266329,
-            745.991986,
-            792.097373,
-            831.921945,
-            865.977814,
-            894.872525,
-            919.22792,
-            939.635932,
-            956.672132,
-            970.827661,
-            982.570665,
-            992.23,
-            1000.0,
-        ]
-    )
-
-    pfull_attrs = {
-        "long_name": "ref full pressure level",
-        "units": "mb",
-        "cartesian_axis": "Z",
-        "positive": "down",
-        "edges": "phalf",
-    }
-
-    phalf_attrs = {
-        "long_name": "ref half pressure level",
-        "units": "mb",
-        "cartesian_axis": "Z",
-        "positive": "down",
-    }
-
-    dset_out = xr.Dataset()
-    dset_out["pfull"] = xr.DataArray(
-        pfull, dims={"pfull": pfull}, coords={"pfull": pfull}, attrs=pfull_attrs
-    )
-    dset_out["phalf"] = xr.DataArray(
-        phalf, dims={"phalf": phalf}, coords={"phalf": phalf}, attrs=phalf_attrs
-    )
-
-    return dset_out
-
-
-def gfdl_plev19_vertical_coord():
-    """Generates GFDL CMIP-like 19-level pressure coordinate
-
-    Returns
-    -------
-    xarray.DataArray
-        GFDL CMIP-like 19-level pressure coordinate
-    """
-
-    plev19 = np.array(
-        [
-            100000.0,
-            92500.0,
-            85000.0,
-            70000.0,
-            60000.0,
-            50000.0,
-            40000.0,
-            30000.0,
-            25000.0,
-            20000.0,
-            15000.0,
-            10000.0,
-            7000.0,
-            5000.0,
-            3000.0,
-            2000.0,
-            1000.0,
-            500.0,
-            100.0,
-        ]
-    )
-
-    plev19_attrs = {
-        "long_name": "pressure",
-        "units": "Pa",
-        "axis": "Z",
-        "positive": "down",
-    }
-
-    dset_out = xr.Dataset()
-    dset_out["plev19"] = xr.DataArray(
-        plev19, dims={"plev19": plev19}, coords={"plev19": plev19}, attrs=plev19_attrs
-    )
-
-    return dset_out
-
-
-def ncar_hybrid_coord():
-    """Generates NCAR CAM2 hybrid vertical coordinate
-
-    Returns
-    -------
-    xarray.DataArray
-        NCAR vertical hybrid coordinate, a, and b coefficients
-    """
-
-    lev = np.array(
-        [
-            2.501651,
-            4.187496,
-            6.66766,
-            10.099201,
-            14.551163,
-            19.943806,
-            26.002806,
-            32.250471,
-            38.050216,
-            42.70557,
-            46.240154,
-            49.511782,
-            53.014888,
-            56.765857,
-            60.782212,
-            65.082732,
-            69.687527,
-            74.618127,
-            79.897583,
-            85.550576,
-            91.603545,
-            98.084766,
-            105.024556,
-            112.455371,
-            120.411924,
-            128.931421,
-            138.053706,
-            147.821421,
-            158.280234,
-            169.479033,
-            181.470176,
-            194.309746,
-            208.057754,
-            222.778457,
-            238.540693,
-            255.418164,
-            273.489746,
-            292.839941,
-            313.559268,
-            335.744541,
-            359.499453,
-            384.935117,
-            412.17043,
-            441.332715,
-            472.55834,
-            505.993262,
-            541.793789,
-            580.127324,
-            621.173066,
-            665.12291,
-            712.182383,
-            762.571445,
-            816.525625,
-            858.699883,
-            886.368125,
-            912.162773,
-            935.873203,
-            957.301758,
-            976.266953,
-            992.556094,
-        ]
-    )
-
-    hyam = np.array(
-        [
-            2.501651e-03,
-            4.187496e-03,
-            6.667660e-03,
-            1.009920e-02,
-            1.455116e-02,
-            1.994381e-02,
-            2.600281e-02,
-            3.225047e-02,
-            3.805022e-02,
-            4.270557e-02,
-            4.624015e-02,
-            4.951178e-02,
-            5.301489e-02,
-            5.676586e-02,
-            6.078221e-02,
-            6.508273e-02,
-            6.968753e-02,
-            7.461813e-02,
-            7.989758e-02,
-            8.555058e-02,
-            9.160354e-02,
-            9.808477e-02,
-            1.050246e-01,
-            1.124554e-01,
-            1.204119e-01,
-            1.289314e-01,
-            1.380537e-01,
-            1.478214e-01,
-            1.582802e-01,
-            1.694790e-01,
-            1.746796e-01,
-            1.726401e-01,
-            1.696388e-01,
-            1.664251e-01,
-            1.629841e-01,
-            1.592996e-01,
-            1.553543e-01,
-            1.511300e-01,
-            1.466068e-01,
-            1.417635e-01,
-            1.365776e-01,
-            1.310247e-01,
-            1.250790e-01,
-            1.187125e-01,
-            1.118957e-01,
-            1.045965e-01,
-            9.678086e-02,
-            8.841227e-02,
-            7.945159e-02,
-            6.985690e-02,
-            5.958334e-02,
-            4.858288e-02,
-            3.680412e-02,
-            2.759706e-02,
-            2.155681e-02,
-            1.592557e-02,
-            1.074934e-02,
-            6.071232e-03,
-            1.930966e-03,
-            -1.648308e-09,
-        ]
-    )
-
-    hybm = np.array(
-        [
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            0.006791,
-            0.02167,
-            0.038419,
-            0.056353,
-            0.075557,
-            0.096119,
-            0.118135,
-            0.14171,
-            0.166953,
-            0.193981,
-            0.222922,
-            0.25391,
-            0.287091,
-            0.32262,
-            0.360663,
-            0.401397,
-            0.445013,
-            0.491715,
-            0.541721,
-            0.595266,
-            0.652599,
-            0.713989,
-            0.779722,
-            0.831103,
-            0.864811,
-            0.896237,
-            0.925124,
-            0.951231,
-            0.974336,
-            0.992556,
-        ]
-    )
-
-    lev_attrs = {
-        "long_name": "hybrid level at midpoints (1000*(A+B))",
-        "units": "level",
-        "positive": "down",
-        "standard_name": "atmosphere_hybrid_sigma_pressure_coordinate",
-        "formula_terms": "a: hyam b: hybm p0: P0 ps: PS",
-    }
-
-    dset_out = xr.Dataset()
-
-    dset_out["hyam"] = xr.DataArray(
-        hyam,
-        dims={"lev": lev},
-        coords={"lev": (lev)},
-        attrs={"long_name": "hybrid A coefficient at layer midpoints"},
-    )
-
-    dset_out["hybm"] = xr.DataArray(
-        hybm,
-        dims={"lev": lev},
-        coords={"lev": (lev)},
-        attrs={"long_name": "hybrid B coefficient at layer midpoints"},
-    )
-
-    dset_out["lev"] = xr.DataArray(
-        lev, dims={"lev": lev}, coords={"lev": (lev)}, attrs=lev_attrs
-    )
-
-    return dset_out
 
 
 def write_to_netcdf(dset_out, outfile, time_dtype="float"):
