+++ mdtf_test_data/tests/test_synthetic_data.py
@@ -8,15 +8,16 @@ import warnings
 
 from mdtf_test_data.synthetic import dataset_stats
 from mdtf_test_data.synthetic import write_to_netcdf
-from mdtf_test_data.synthetic import ncar_hybrid_coord
-from mdtf_test_data.synthetic import gfdl_plev19_vertical_coord
-from mdtf_test_data.synthetic import gfdl_vertical_coord
 from mdtf_test_data.synthetic import generate_synthetic_dataset
 
 from mdtf_test_data.synthetic.time import generate_daily_time_axis
 from mdtf_test_data.synthetic.time import generate_hourly_time_axis
 from mdtf_test_data.synthetic.time import generate_monthly_time_axis
 from mdtf_test_data.synthetic.time import xr_times_from_tuples
+
+from mdtf_test_data.synthetic.vertical import gfdl_plev19_vertical_coord
+from mdtf_test_data.synthetic.vertical import gfdl_vertical_coord
+from mdtf_test_data.synthetic.vertical import ncar_hybrid_coord
 
 __all__ = [
     "test_xr_times_from_tuples_ncar",
